Issue #3: Handle IA urls in overwrite mode
This plugin targets the IA workflow used with the American Numismatic Society.
IA usually appends information on the CSV they provide, so it is desirable to get the IA specific links rewritten so there are no duplicates.
Also, non-IA links should always be preserved.
This should be done in a way that appending/overwritting can be chosen from the UI if requested by users.

Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>